03-05 [HTTP 상태 처리하기] POST,DELETE,PUT에 대한 ResponseEntity 응답

+++ chapter03/src/main/kotlin/com/microservices/chapter03/CustomerController.kt
@@ -34,18 +34,35 @@ class CustomerController {
     }
 
     @PostMapping("/customer/")
-    fun createCustomer(@RequestBody customer: Customer) {
+    fun createCustomer(@RequestBody customer: Customer): ResponseEntity<Unit> {
         customerService.createCustomer(customer)
+
+        return ResponseEntity(Unit, HttpStatus.CREATED)
     }
 
     @DeleteMapping("/customer/{id}")
-    fun deleteCustomer(@PathVariable id: Int) {
-        customerService.deleteCustomer(id)
+    fun deleteCustomer(@PathVariable id: Int): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (customerService.getCustomer(id) != null) {
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping("/customer/{id}")
-    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer) {
-        customerService.updateCustomer(id, customer)
+    fun updateCustomer(@PathVariable id: Int, 
+                       @RequestBody customer: Customer): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if (customerService.getCustomer(id) != null) {
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.ACCEPTED
+        }
+
+        return ResponseEntity(Unit, status)
     }
   
 }
